# application/services/custom_services/formservices/src/models/entitymodels/forms.ts
export interface forms 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   email: { type: String },
   dob: { type: String }
}